+++ tests_zemu/try_deploy.mjs
@@ -1,0 +1,20 @@
+import TransportNodeHid from '@ledgerhq/hw-transport-node-hid'
+import CasperApp from '@zondax/ledger-casper'
+
+// undelegate__type_by_hash__payment_system
+const txBlobStrNormal = "0001ca93ac1705187071d67b83c7ff0efe8108e8ec4530575d7726879333dbdabe7ca087c0377901000080ee360000000000020000000000000077acc4d0e0b59fd05c8b0210d10722c6d00381f8059cb2a8c1a513e8acd004550a0000000000000000000000000000000000000000000000000000000000000000000000010101010101010101010101010101010101010101010101010101010101010102020202020202020202020202020202020202020202020202020202020202020303030303030303030303030303030303030303030303030303030303030303040404040404040404040404040404040404040404040404040404040404040405050505050505050505050505050505050505050505050505050505050505050606060606060606060606060606060606060606060606060606060606060606070707070707070707070707070707070707070707070707070707070707070708080808080808080808080808080808080808080808080808080808080808080909090909090909090909090909090909090909090909090909090909090909070000006d61696e6e6574ca815f7a03a7066e117cf30b9b141b2d73e5a7f2448c43d0466126b8e8a2355e00000000000100000006000000616d6f756e74050000000400ca9a3b0800000000002800000003000000553634080000000000000000000000050300000055333204000000ffffffff040400000055526566210000004acfcf6c684c58caf6b3296e3a97c4a04afaf77bb875ca9a40a45db254e94a75010c020000005538010000000b03030000004b6579210000000401010101010101010101010101010101010101010101010101010101010101010b030000004b6579210000000701010101010101010101010101010101010101010101010101010101010101010b03000000493634080000000000000000000080020a0000004f7074696f6e285538290200000001640d030400000055526566210000004acfcf6c684c58caf6b3296e3a97c4a04afaf77bb875ca9a40a45db254e94a75020c030000005533320400000000000000041a0000005475706c6533285b55382c20426f6f6c2c20537472696e675d290c0000000001060000007475706c65331403000a030000004b6579210000000c01010101010101010101010101010101010101010101010101010101010101010b04000000553132381100000010ffffffffffffffffffffffffffffffff06030000004b6579210000000601010101010101010101010101010101010101010101010101010101010101010b0400000055526566210000004acfcf6c684c58caf6b3296e3a97c4a04afaf77bb875ca9a40a45db254e94a75000c0e0000006c6973742d7075626c69636b657904000000000000000e16030000004b6579210000000001010101010101010101010101010101010101010101010101010101010101010b1d000000526573756c74207b206f6b3a20426f6f6c2c206572723a20493332207d0200000001001000010300000049333204000000ffffff7f0104000000553235362100000020ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff07030000004b657922000000020101010101010101010101010101010101010101010101010101010101010101070b0a0000004f7074696f6e2855382901000000000d030400000055526566210000004acfcf6c684c58caf6b3296e3a97c4a04afaf77bb875ca9a40a45db254e94a75070c0900000062797465617272617904000000000000000e03110000005475706c6532285b55382c205536345d29090000000b570400000000000013030506000000537472696e67110000000d00000073616d706c652d737472696e670a03000000493634080000000000000000000000020300000049363408000000ffffffffffffff7f02030000004b6579210000000301010101010101010101010101010101010101010101010101010101010101010b0400000055526566210000004acfcf6c684c58caf6b3296e3a97c4a04afaf77bb875ca9a40a45db254e94a75040c09000000627974656172726179240000002000000001010101010101010101010101010101010101010101010101010101010101010e03030000004b6579210000000a00000000000000000000000000000000000000000000000000000000000000000b090000005075626c69634b65790100000000160300000055363408000000ffffffffffffffff050400000055526566210000004acfcf6c684c58caf6b3296e3a97c4a04afaf77bb875ca9a40a45db254e94a75050c030000004b6579210000000801010101010101010101010101010101010101010101010101010101010101010b0400000055323536010000000007030000004b6579090000000500000000000000000b04000000426f6f6c01000000000002000000553801000000ff030a000000011398f62c6d1a457c51ba6a4b5f3dbd2f69fca93216218dc8997e416bd17d93ca014c38e7abda8af0b36035ff33b281db799f1986343b217fe76afb116a0bd5a90f0983d875d49b32eedf7608dbea42189f8a360deb4a6622a81b51e052a6896301013b6a27bcceb6a42d62a3a8d02a6f0d73653215771de243a63ac048a18b59da290176b2189432e2b42be4f9a757b8d5a65dced3fdab364c3aaddb8b68ca87125f0df148b6e1909db5f0c68527ced040c21711bcaccf206177bd57d5ec8d23e8cd01018139770ea87d175f56a35466c34c7ecccb8d8a91b4ee37a25df60f5b8fc9b39401ccd95d3b27e49ef6dcbd28e6017069954df1fea1a7389eb38f4d2e30aeac649e0339cb173376ae26a4b579059f89c029e56e95a90cceaa475096e9f9ef3dfd0b018a875fff1eb38451577acd5afee405456568dd7c89e090863a0557bc7af49f170166b19f855c6ba395727dfd822274d2a3707a7602daf341c9ad8e8222b298247f57f0715393f938d34f8bbdb5a57ab28efcbd19313540b41e4b43e35d953dbd0501ca93ac1705187071d67b83c7ff0efe8108e8ec4530575d7726879333dbdabe7c019763175645b99d559ab84dc7f46adfbbc5bdf3a59e0e006d286d2563d3683026acdd9cfe3779f71be271b4da4eb1364b253004f15a10d2774734a185a49da70f0202531fe6068134503d2723133227c867ac8fa6c83c537e9a44c3c5bdbdcb1fe33702dffb7ae96b938c21b3b88d656d6a1396105a6f718e85b5a26a66327151fe6a1c3861744ee3817ff10073478ab506b4391f84ba044d34c0f2a0b098fae9d13d92020256b328b30c8bf5839e24058747879408bdb36241dc9c2e7c619faa12b2920967026b921736e6acb062ad798173be50cc4637b31a399091002b20df6237de8b30ca42e5a2d0e920016b8468473130ec3b54b01da49cc024ef5536df7263c4f4b3770202989c0b76cb563971fdc9bef31ec06c3560f3249d6ee9e5d83c57625596e05f6f02123e6aa43180b84cb53e215ff8aa3f54c1c58424d22cbef4b5b7578391bde1bb4827eb92bfc9555cba43c700ef769633928150c5af07772848e4f356ce07170202031b84c5567b126440995d3ed5aaba0565d71e1834604819ff9c17f5e9d5dd078f02b50fc36b8fe516c5cb7d6178b656cd4b4f7436a4f56c23e0af62d89be930e0d256ed930fd826a1c28fcb8d83c432d5c095e861589da593c88574b76e5e91ceec020362c0a046dacce86ddd0343c6d3c7c79c2208ba0d9c9cf24a6d046d21d21f90f702749c6bd64ed493e20476e713383555fe3faffe300d5c783611321620a1dfb38d2abb0dd0834fdc40258c6f4afaf9c352db0b4570eeeb9faaed592c3adee3b3f4"
+
+async function main() {
+  const transport = await TransportNodeHid.default.open();
+
+  const app = new CasperApp.default(transport);
+  const PATH = "m/44'/506'/0'/0/0"
+
+  const txBlob = Buffer.from(txBlobStrNormal, "hex");
+  const respRequest = await app.sign(PATH, txBlob);
+  console.log(respRequest)
+}
+
+; (async () => {
+  await main()
+})()
