+++ tests_zemu/try_txnV1.mjs
@@ -1,0 +1,20 @@
+import TransportNodeHid from '@ledgerhq/hw-transport-node-hid'
+import CasperApp from '@zondax/ledger-casper'
+
+const txBlobStrNormal =
+"0103000000000000000000010020000000020091010000f801000097fb1275cf8606bcb39493d29f7c7af4e4ed3ce1e418c2e33cb5d913959390cb0600000000000000000001003700000002003f00000003004700000004005200000005006f000000450100000200000000000000000001000100000023000000000202531fe6068134503d2723133227c867ac8fa6c83c537e9a44c3c5bdbdcb1fe337a087c0377901000080ee360000000000070000006d61696e6e657403000000000000000000010001000000020002000000030000000164010400000000008d000000000400000006000000616d6f756e74010000000008020000006964090000000100000000000000000d0506000000736f75726365210000004acfcf6c684c58caf6b3296e3a97c4a04afaf77bb875ca9a40a45db254e94a75010c0600000074617267657420000000ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff0f2000000001000f00000001000000000000000000010000000002000f00000001000000000000000000010000000203000f000000010000000000000000000100000000010000000202531fe6068134503d2723133227c867ac8fa6c83c537e9a44c3c5bdbdcb1fe33702f8918d6446914562ac3e176b2413bd479d367069a35f2ea7ff930f027d34615f13fa828b8dc3a6df56afc817f59c1e478120e298bf5a17024b658f470c496104";
+
+async function main() {
+  const transport = await TransportNodeHid.default.open();
+
+  const app = new CasperApp.default(transport);
+  const PATH = "m/44'/506'/0'/0/0"
+
+  const txBlob = Buffer.from(txBlobStrNormal, "hex");
+  const respRequest = await app.sign(PATH, txBlob);
+  console.log(respRequest)
+}
+
+; (async () => {
+  await main()
+})()
